+++ lib/src/failure_or_result.dart
@@ -8,12 +8,14 @@ class FailureOrResult<T> {
     required String code,
     required String message,
     Map<String, dynamic>? details,
+    StackTrace? stackTrace,
     FailureType type = FailureType.exception,
   })  : failure = Failure(
           code: code,
           message: message,
-          type: type,
           details: details,
+          stackTrace: stackTrace,
+          type: type,
         ),
         result = null;
 
@@ -23,14 +25,13 @@ class FailureOrResult<T> {
 
   bool get wasSuccessful => failure == null;
 
-  FailureOrResult<T2> map<T2>(T2 Function(T result) convert) {
-    if (wasSuccessful) {
-      return FailureOrResult.success(convert(result as T));
-    } else {
-      return FailureOrResult.failure(
-        code: failure!.code,
-        message: failure!.message,
-      );
-    }
-  }
+  FailureOrResult<T2> map<T2>(T2 Function(T result) convert) => wasSuccessful
+      ? FailureOrResult.success(convert(result as T))
+      : FailureOrResult.failure(
+          code: failure!.code,
+          message: failure!.message,
+          details: failure!.details,
+          stackTrace: failure!.stackTrace,
+          type: failure!.type,
+        );
 }