+++ lib/src/state.dart
@@ -1,5 +1,7 @@
 import 'package:clean_redux/src/action.dart';
+import 'package:clean_redux/src/failure.dart';
 import 'package:clean_redux/src/reducer.dart';
+import 'package:clean_redux/src/use_case.dart';
 
 /// An abstract class representing the state of a Redux store.
 ///
@@ -16,4 +18,49 @@ abstract class State<T extends State<T>> {
 
   /// A method that creates a copy of the current state.
   T copyWith();
+}
+
+// Generic state to keep info about failures
+class FailuresState extends State<FailuresState> {
+  final Map<Type, Failure?> failures;
+
+  FailuresState({
+    required this.failures,
+  }) : super(
+          FailuresState.updateFailure.reducer,
+        );
+
+  FailuresState.initial() : this(failures: {});
+
+  Failure? getFailure<T extends UseCase>() => failures[FailedAction<T>];
+
+  factory FailuresState.updateFailure(
+    FailuresState state,
+    FailedAction<UseCase> action,
+  ) =>
+      state.copyWith(
+        failures: {
+          ...state.failures,
+          action.runtimeType: action.failure,
+        },
+      );
+
+  factory FailuresState.resetFailure(
+    FailuresState state,
+    ResetFailureAction<UseCase> action,
+  ) =>
+      state.copyWith(
+        failures: {
+          ...state.failures,
+          action.runtimeType: null,
+        },
+      );
+
+  @override
+  FailuresState copyWith({
+    Map<Type, Failure?>? failures,
+  }) =>
+      FailuresState(
+        failures: failures ?? this.failures,
+      );
 }